+++ architect_py/grpc/client.py
@@ -17,6 +17,8 @@ class GrpcClient:
     endpoint: str
     channel: grpc.aio.Channel
     jwt: str | None = None
+    as_user: str | None = None
+    as_role: str | None = None
 
     def __init__(
         self,
@@ -25,9 +27,13 @@ class GrpcClient:
         port: int,
         use_ssl: bool = False,
         options: Sequence[Tuple[str, Any]] | None = None,
+        as_user: str | None = None,
+        as_role: str | None = None,
     ):
         scheme = "https" if use_ssl else "http"
         self.endpoint = f"{scheme}://{host}:{port}"
+        self.as_user = as_user
+        self.as_role = as_role
         if use_ssl:
             credentials = grpc.ssl_channel_credentials()
             self.channel = grpc.aio.secure_channel(
@@ -38,6 +44,16 @@ class GrpcClient:
 
     def set_jwt(self, jwt: str | None):
         self.jwt = jwt
+
+    def metadata(self) -> Sequence[Tuple[str, str]]:
+        metadata = []
+        if self.jwt is not None:
+            metadata.append(("authorization", f"Bearer {self.jwt}"))
+        if self.as_user is not None:
+            metadata.append(("x-architect-user", self.as_user))
+        if self.as_role is not None:
+            metadata.append(("x-architect-role", self.as_role))
+        return metadata
 
     async def close(self):
         await self.channel.close()
@@ -75,10 +91,7 @@ class GrpcClient:
             request_serializer=encoder.encode,
             response_deserializer=decoder.decode,
         )
-        if self.jwt is not None:
-            metadata = (("authorization", f"Bearer {self.jwt}"),)
-        else:
-            metadata = ()
+        metadata = self.metadata()
         return await stub(request, metadata=metadata)
 
     async def unary_stream(
@@ -100,10 +113,7 @@ class GrpcClient:
             request_serializer=encoder.encode,
             response_deserializer=decoder.decode,
         )
-        if self.jwt is not None:
-            metadata = (("authorization", f"Bearer {self.jwt}"),)
-        else:
-            metadata = ()
+        metadata = self.metadata()
         call: grpc.aio._base_call.UnaryStreamCall[
             RequestType[ResponseTypeGeneric], ResponseTypeGeneric
         ] = stub(request, metadata=metadata)